[update]MVCモデルの分割を決定
やっぱりモデルは分割した。
README.md: 環境構築項目を作成
