+++ model/modelope.go
@@ -1,5 +1,3 @@
 package model
 
 // テーブルに対する処理
-
-// insert record
